+++ src/org/yccheok/jstock/analysis/StockRelativeHistoryOperator.java
@@ -144,7 +144,8 @@ public class StockRelativeHistoryOperator extends AbstractOperator {
         java.util.List<Double> highs = new java.util.ArrayList<Double>();
         java.util.List<Double> lows = new java.util.ArrayList<Double>();
         java.util.List<Double> closes = new java.util.ArrayList<Double>();
-        java.util.List<Integer> volumes = new java.util.ArrayList<Integer>();
+        // TODO: CRITICAL LONG BUG REVISED NEEDED.
+        java.util.List<Long> volumes = new java.util.ArrayList<Long>();
 
         final int numOfCalendar = stockHistoryServer.getNumOfCalendar();
         /* Fill up stocks. */
